requirements.txt
Django>=3.0.7,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0
flake8>=3.8.2,<3.9.0
Pillow>=5.3.0,<5.4.0
django-webpack-loader>=0.6.0,<0.7.0



